Remove add & remove minimum order group handlers/translators.
Cover them by `changeGroup` to reduce the amount of duplicated code.

Address https://github.com/woocommerce/google-listings-and-ads/pull/1410#discussion_r871578823
