+++ js/src/components/shipping-rate-section/minimum-order-card/changeGroup.js
@@ -7,12 +7,13 @@
  * Returns new shipping rates value, with updated group.
  *
  * Will set `shipingRate.options.free_shipping_threshold` to `newGroup.threshold`
- * if the shipping rate country is part of `newGroup.countries`,
- * and set `threshold` to `undefined` for the countries which were removed from the group.
+ * for the shipping rate countries that are listed in `newGroup.countries` (if provided),
+ * and set `threshold` to `undefined` for the countries which were in `oldGroup.countries` (if provided),
+ * but not in the new.
  *
  * @param {Array<ShippingRate>} value Shipping rates.
- * @param {MinimumOrderGroup} oldGroup Old minimum order group.
- * @param {MinimumOrderGroup} newGroup New minimum order group.
+ * @param {MinimumOrderGroup} [oldGroup] Old minimum order group.
+ * @param {MinimumOrderGroup} [newGroup] New minimum order group.
  */
 export const changeGroup = ( value, oldGroup, newGroup ) => {
 	return value.map( ( shippingRate ) => {
@@ -23,14 +24,20 @@ export const changeGroup = ( value, oldGroup, newGroup ) => {
 			},
 		};
 
-		if ( newGroup.countries.includes( newShippingRate.country ) ) {
+		if (
+			newGroup &&
+			newGroup.countries.includes( newShippingRate.country )
+		) {
 			/**
 			 * Shipping rate's country exists in the new value countries,
 			 * so we just assign the new value threshold.
 			 */
 			newShippingRate.options.free_shipping_threshold =
 				newGroup.threshold;
-		} else if ( oldGroup.countries.includes( newShippingRate.country ) ) {
+		} else if (
+			oldGroup &&
+			oldGroup.countries.includes( newShippingRate.country )
+		) {
 			/**
 			 * Shipping rate's country does not exist in the new value countries,
 			 * but it exists in the old value countries.
