--- js/src/components/shipping-rate-section/minimum-order-card/handlers/getMinimumOrderAddHandler.test.js
@@ -1,78 +1,0 @@
-/**
- * Internal dependencies
- */
-import { addGroup } from './getMinimumOrderAddHandler';
-
-describe( 'addGroup( value, newGroup )', () => {
-	const value = Object.freeze( [
-		{
-			id: '1',
-			country: 'US',
-			method: 'flat_rate',
-			currency: 'USD',
-			rate: 20,
-			options: {},
-		},
-		{
-			id: '2',
-			country: 'AU',
-			method: 'flat_rate',
-			currency: 'USD',
-			rate: 20,
-			options: {
-				free_shipping_threshold: 50,
-			},
-		},
-		{
-			id: '3',
-			country: 'CN',
-			method: 'flat_rate',
-			currency: 'USD',
-			rate: 25,
-			options: {
-				free_shipping_threshold: 50,
-			},
-		},
-	] );
-
-	it( 'returns a new value (`shippingRate.options.free_shipping_threshold` set to `newGroup.threshold`)', () => {
-		const newGroup = {
-			countries: [ 'US' ],
-			currency: 'USD',
-			threshold: 30,
-		};
-
-		expect( addGroup( value, newGroup ) ).toStrictEqual( [
-			{
-				id: '1',
-				country: 'US',
-				method: 'flat_rate',
-				currency: 'USD',
-				rate: 20,
-				options: {
-					free_shipping_threshold: 30,
-				},
-			},
-			{
-				id: '2',
-				country: 'AU',
-				method: 'flat_rate',
-				currency: 'USD',
-				rate: 20,
-				options: {
-					free_shipping_threshold: 50,
-				},
-			},
-			{
-				id: '3',
-				country: 'CN',
-				method: 'flat_rate',
-				currency: 'USD',
-				rate: 25,
-				options: {
-					free_shipping_threshold: 50,
-				},
-			},
-		] );
-	} );
-} );
