--- js/src/components/shipping-rate-section/minimum-order-card/handlers/index.js
@@ -1,3 +1,0 @@
-export { addGroup } from './getMinimumOrderAddHandler';
-export { changeGroup } from './getMinimumOrderChangeHandler';
-export { deleteGroup } from './getMinimumOrderDeleteHandler';
